refactor: merge chat-ai-tools and chat-core into single chat-ai package
- Merge @repo/chat-ai-tools and @repo/chat-core into @repo/chat-ai
- Resolve circular dependency issues between packages
- Consolidate types and tools in single package for better maintainability
- Update all imports and dependencies across the codebase
- Remove redundant packages and simplify package structure
- Maintain proper type inference using InferUITools from AI SDK
- Update db/chat to reference new consolidated package
