+++ apps/chat/package.json
@@ -28,7 +28,7 @@
     "@clerk/nextjs": "catalog:",
     "@db/chat": "workspace:*",
     "@repo/ai": "workspace:*",
-    "@repo/chat-core": "workspace:*",
+    "@repo/chat-ai": "workspace:*",
     "@repo/chat-services": "workspace:*",
     "@repo/chat-trpc": "workspace:*",
     "@repo/lib": "workspace:^",
