+++ apps/chat/src/app/(chat)/_components/chat-interface.tsx
@@ -39,7 +39,7 @@ import { useBillingContext } from "~/hooks/use-billing-context";
 import { ChatErrorHandler } from "~/lib/errors/chat-error-handler";
 import { ChatErrorType } from "~/lib/errors/types";
 import type { ChatError } from "~/lib/errors/types";
-import type { LightfastAppChatUIMessage } from "@repo/chat-core/types";
+import type { LightfastAppChatUIMessage } from "@repo/chat-ai/types";
 import type { ChatRouterOutputs } from "@api/chat";
 import type { ArtifactApiResponse } from "~/components/artifacts/types";
 import { useTRPC } from "@repo/chat-trpc/react";
