+++ apps/chat/src/hooks/use-data-stream.tsx
@@ -2,7 +2,7 @@
 
 import React, { createContext, useContext, useMemo, useState } from 'react';
 import type { DataUIPart } from 'ai';
-import type { LightfastAppChatUICustomDataTypes } from '@repo/chat-core/types';
+import type { LightfastAppChatUICustomDataTypes } from '@repo/chat-ai/types';
 
 interface DataStreamContextValue {
   dataStream: DataUIPart<LightfastAppChatUICustomDataTypes>[];
