--- packages/chat-ai-tools/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@repo/chat-ai-tools",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "exports": {
-    "./tools": {
-      "default": "./src/tools.ts"
-    }
-  },
-  "scripts": {
-    "build": "tsc",
-    "clean": "git clean -xdf .cache .turbo dist node_modules",
-    "dev": "tsc --watch",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "ai": "catalog:"
-  },
-  "devDependencies": {
-    "@repo/eslint-config": "workspace:*",
-    "@repo/prettier-config": "workspace:*",
-    "@repo/typescript-config": "workspace:*",
-    "@types/node": "catalog:",
-    "eslint": "catalog:",
-    "prettier": "catalog:",
-    "typescript": "catalog:"
-  },
-  "prettier": "@repo/prettier-config"
-}
