--- packages/chat-ai-tools/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "@repo/typescript-config/internal-package.json",
-  "compilerOptions": {
-    "types": ["node"],
-    "moduleResolution": "bundler",
-    "module": "ESNext",
-    "target": "ES2020",
-    "rootDir": "src",
-    "outDir": "dist"
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist"]
-}
